Deepen stack request handling
